# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TransitOps API",
    "description": "Fleet operations platform \u2014 manage vehicles, drivers, trips, maintenance, fuel, expenses, and reports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/trips": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "List trips",
        "description": "Returns all trips. Drivers only see their own assigned trips.",
        "operationId": "list_trips_trips_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TripResponse"
                  },
                  "type": "array",
                  "title": "Response List Trips Trips Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Create a new trip (Draft)",
        "description": "Create a trip in Draft status. Validates Rules 2-5 at creation. Requires fleet_manager or driver role.",
        "operationId": "create_trip_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business rule violation (Rules 2-5)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Vehicle or driver not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/trips/{trip_id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get trip by ID",
        "description": "Returns a single trip.",
        "operationId": "get_trip_trips__trip_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Trip not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trips/{trip_id}/dispatch": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Dispatch a trip",
        "description": "Transition Draft \u2192 Dispatched. Re-validates Rules 2, 3, 4 (race-condition guard). Rule 6: sets vehicle + driver to On Trip.",
        "operationId": "dispatch_trip_trips__trip_id__dispatch_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business rule violation (Rules 2-4)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Trip not found"
          },
          "409": {
            "description": "Trip not in Draft status"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trips/{trip_id}/complete": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Complete a trip",
        "description": "Transition Dispatched \u2192 Completed. Rule 7: restores vehicle + driver to Available. Design decision #1: auto-creates FuelLog. Design decision #2: updates Vehicle.odometer_km.",
        "operationId": "complete_trip_trips__trip_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripComplete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Trip not found"
          },
          "409": {
            "description": "Trip not in Dispatched status"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trips/{trip_id}/cancel": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Cancel a trip",
        "description": "Cancel a Draft or Dispatched trip. Design decision #3: Draft cancellation has no side effects. Dispatched cancellation restores vehicle + driver to Available (Rule 8). Completed/Cancelled trips cannot be cancelled (409).",
        "operationId": "cancel_trip_trips__trip_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Trip not found"
          },
          "409": {
            "description": "Trip cannot be cancelled from current status"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripComplete": {
        "properties": {
          "actual_distance_km": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Actual Distance Km",
            "examples": [
              345.0
            ]
          },
          "fuel_consumed_liters": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Fuel Consumed Liters",
            "examples": [
              42.5
            ]
          },
          "final_odometer_km": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Odometer Km",
            "description": "If provided, sets vehicle odometer to this value. Otherwise, odometer is incremented by actual_distance_km.",
            "examples": [
              15345.0
            ]
          }
        },
        "type": "object",
        "required": [
          "actual_distance_km",
          "fuel_consumed_liters"
        ],
        "title": "TripComplete",
        "description": "Payload for completing a trip \u2014 design decisions #1 and #2."
      },
      "TripCreate": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "examples": [
              "Warehouse A"
            ]
          },
          "destination": {
            "type": "string",
            "title": "Destination",
            "examples": [
              "Distribution Center B"
            ]
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "examples": [
              "uuid-of-vehicle"
            ]
          },
          "driver_id": {
            "type": "string",
            "title": "Driver Id",
            "examples": [
              "uuid-of-driver"
            ]
          },
          "cargo_weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cargo Weight Kg",
            "examples": [
              1800.0
            ]
          },
          "planned_distance_km": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Planned Distance Km",
            "examples": [
              350.0
            ]
          },
          "revenue": {
            "type": "number",
            "minimum": 0.0,
            "title": "Revenue",
            "default": 0.0,
            "examples": [
              2500.0
            ]
          }
        },
        "type": "object",
        "required": [
          "source",
          "destination",
          "vehicle_id",
          "driver_id",
          "cargo_weight_kg",
          "planned_distance_km"
        ],
        "title": "TripCreate"
      },
      "TripResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "cargo_weight_kg": {
            "type": "number",
            "title": "Cargo Weight Kg"
          },
          "planned_distance_km": {
            "type": "number",
            "title": "Planned Distance Km"
          },
          "actual_distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Distance Km"
          },
          "fuel_consumed_liters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Consumed Liters"
          },
          "revenue": {
            "type": "number",
            "title": "Revenue"
          },
          "status": {
            "$ref": "#/components/schemas/TripStatus"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "dispatched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatched At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "destination",
          "vehicle_id",
          "driver_id",
          "cargo_weight_kg",
          "planned_distance_km",
          "actual_distance_km",
          "fuel_consumed_liters",
          "revenue",
          "status",
          "created_by",
          "dispatched_at",
          "completed_at",
          "cancelled_at",
          "created_at"
        ],
        "title": "TripResponse"
      },
      "TripStatus": {
        "type": "string",
        "enum": [
          "Draft",
          "Dispatched",
          "Completed",
          "Cancelled"
        ],
        "title": "TripStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}